Using a secure RNG from sys/random.h
Removed the dependence on time() for seeding. Implemented getrandom() for each call of rand() to prevent duplicate seeds.
Previous modulus operations reduced the cycle length for determining indices so instead of relying on rand() for
when a character is capatilized, the list of available characters now includes all alphabetically capitalized ASCII characters.
This commit should decrease the number of repeated substrings and allow for large batch generation of passwords. A sample output
file has been included to pwGen/Main and README reflects some of these changes.